# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AuraLearning Elite API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /**
     * @minimum 6
     * @maximum 10
     */
  classLevel: number;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Student {
  id: number;
  name: string;
  email: string;
  classLevel: number;
  points: number;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  student: Student;
  token: string;
}

export interface Subject {
  id: number;
  name: string;
  classLevel: number;
  icon: string;
  color: string;
  totalChapters: number;
  completedChapters: number;
  totalTopics: number;
  completedTopics: number;
}

export type ChapterStatus = typeof ChapterStatus[keyof typeof ChapterStatus];


export const ChapterStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  complete: 'complete',
  marked_for_review: 'marked_for_review',
} as const;

export type TopicStatus = typeof TopicStatus[keyof typeof TopicStatus];


export const TopicStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  complete: 'complete',
  needs_review: 'needs_review',
} as const;

export interface Topic {
  id: number;
  chapterId: number;
  title: string;
  status: TopicStatus;
  isImportant?: boolean;
}

export interface Chapter {
  id: number;
  subjectId: number;
  number: number;
  title: string;
  status: ChapterStatus;
  mastery: number;
  topics?: Topic[];
}

export interface SubjectDetail {
  id: number;
  name: string;
  classLevel: number;
  icon: string;
  color: string;
  mastery: number;
  weeklyGoal: number;
  weeklyCompleted: number;
  chapters: Chapter[];
}

export interface Progress {
  id: number;
  studentId: number;
  subjectId: number;
  chapterId: number;
  /** @nullable */
  topicId: number | null;
  status: string;
  updatedAt: string;
}

export type ProgressInputStatus = typeof ProgressInputStatus[keyof typeof ProgressInputStatus];


export const ProgressInputStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  complete: 'complete',
  needs_review: 'needs_review',
} as const;

export interface ProgressInput {
  topicId: number;
  status: ProgressInputStatus;
}

export type TestDifficulty = typeof TestDifficulty[keyof typeof TestDifficulty];


export const TestDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type TestLevel = typeof TestLevel[keyof typeof TestLevel];


export const TestLevel = {
  starting: 'starting',
  mid: 'mid',
  complete: 'complete',
  mastering: 'mastering',
  mastered: 'mastered',
} as const;

export interface Test {
  id: number;
  title: string;
  subjectId: number;
  subjectName: string;
  classLevel: number;
  totalQuestions: number;
  duration: number;
  difficulty: TestDifficulty;
  level: TestLevel;
  /** @nullable */
  chapterName?: string | null;
}

export type QuestionType = typeof QuestionType[keyof typeof QuestionType];


export const QuestionType = {
  multiple_choice: 'multiple_choice',
  fill_blank: 'fill_blank',
  short_answer: 'short_answer',
  long_answer: 'long_answer',
  match: 'match',
} as const;

export interface Question {
  id: number;
  text: string;
  type: QuestionType;
  options: string[];
  /** @nullable */
  correctOption?: number | null;
  /** @nullable */
  explanation?: string | null;
}

export interface TestDetail {
  id: number;
  title: string;
  subjectId: number;
  subjectName: string;
  classLevel: number;
  totalQuestions: number;
  duration: number;
  difficulty: string;
  level: string;
  /** @nullable */
  chapterName?: string | null;
  questions: Question[];
}

export interface Answer {
  questionId: number;
  /** @nullable */
  selectedOption?: number | null;
  /** @nullable */
  textAnswer?: string | null;
}

export interface TestSubmission {
  answers: Answer[];
}

export interface WrongAnswer {
  questionId: number;
  questionText: string;
  yourAnswer: string;
  correctAnswer: string;
}

export interface TopicScore {
  topic: string;
  score: number;
}

export interface TestResult {
  id: number;
  testId: number;
  testTitle: string;
  subjectName: string;
  classLevel: number;
  totalQuestions: number;
  attempted: number;
  correct: number;
  incorrect: number;
  score: number;
  mastery: number;
  completedAt: string;
  wrongAnswers?: WrongAnswer[];
  topicBreakdown?: TopicScore[];
}

export type ImportantQuestionDifficulty = typeof ImportantQuestionDifficulty[keyof typeof ImportantQuestionDifficulty];


export const ImportantQuestionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface ImportantQuestion {
  id: number;
  text: string;
  subject: string;
  chapter: string;
  difficulty: ImportantQuestionDifficulty;
  frequency: string;
  solution: string;
  tags: string[];
  /** @nullable */
  year?: string | null;
  isBookmarked?: boolean;
}

export interface ImportantQuestionList {
  questions: ImportantQuestion[];
  total: number;
  page: number;
  limit: number;
}

export type BookmarkType = typeof BookmarkType[keyof typeof BookmarkType];


export const BookmarkType = {
  question: 'question',
  topic: 'topic',
  chapter: 'chapter',
} as const;

export interface Bookmark {
  id: number;
  type: BookmarkType;
  referenceId: number;
  title: string;
  subject: string;
  chapter: string;
  createdAt: string;
}

export type BookmarkInputType = typeof BookmarkInputType[keyof typeof BookmarkInputType];


export const BookmarkInputType = {
  question: 'question',
  topic: 'topic',
  chapter: 'chapter',
} as const;

export interface BookmarkInput {
  type: BookmarkInputType;
  referenceId: number;
  title: string;
  subject: string;
  chapter: string;
}

export interface DashboardSummary {
  student: Student;
  totalPoints: number;
  weeklyGoalCompleted: number;
  weeklyGoalTotal: number;
  overallMastery: number;
  totalTestsTaken: number;
  averageScore: number;
  currentChapter: Chapter;
  streakDays: number;
}

export interface SubjectProgressSummary {
  subjectId: number;
  subjectName: string;
  color: string;
  icon: string;
  completedTopics: number;
  totalTopics: number;
  mastery: number;
}

export type ListSubjectsParams = {
classLevel?: number;
};

export type ListTopicsParams = {
chapterId?: number;
};

export type ListTestsParams = {
classLevel?: number;
subjectId?: number;
};

export type ListImportantQuestionsParams = {
classLevel?: number;
subject?: string;
page?: number;
limit?: number;
};