# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AuraLearning Elite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Unauthorized
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects for a class
      parameters:
        - name: classLevel
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get subject with chapters
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectDetail"
  /subjects/{subjectId}/chapters:
    get:
      operationId: listChapters
      tags: [subjects]
      summary: List chapters for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
  /topics:
    get:
      operationId: listTopics
      tags: [subjects]
      summary: List topics for a chapter
      parameters:
        - name: chapterId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
...
  /progress/topic:
    post:
      operationId: updateTopicProgress
      tags: [progress]
      summary: Mark topic as complete or in-progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
...
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: List bookmarks for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"

    post:
      operationId: addBookmark
      tags: [bookmarks]
      summary: Add bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{bookmarkId}:
    delete:
      operationId: removeBookmark
      tags: [bookmarks]
      summary: Remove bookmark
      parameters:
        - name: bookmarkId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /results/{resultId}:
    get:
      operationId: getResult
      tags: [results]
      summary: Get detailed test result
      parameters:
        - name: resultId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, classLevel]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        classLevel:
          type: integer
          minimum: 6
          maximum: 10
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [student, token]
      properties:
        student:
          $ref: "#/components/schemas/Student"
        token:
          type: string
    Student:
      type: object
      required: [id, name, email, classLevel, points, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        classLevel:
          type: integer
        points:
          type: integer
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    Subject:
      type: object
      required: [id, name, classLevel, icon, color, totalChapters, completedChapters, totalTopics, completedTopics]
      properties:
        id:
          type: integer
        name:
          type: string
        classLevel:
          type: integer
        icon:
          type: string
        color:
          type: string
        totalChapters:
          type: integer
        completedChapters:
          type: integer
        totalTopics:
          type: integer
        completedTopics:
          type: integer
    SubjectDetail:
      type: object
      required: [id, name, classLevel, icon, color, chapters, mastery, weeklyGoal, weeklyCompleted]
      properties:
        id:
          type: integer
        name:
          type: string
        classLevel:
          type: integer
        icon:
          type: string
        color:
          type: string
        mastery:
          type: integer
        weeklyGoal:
          type: integer
        weeklyCompleted:
          type: integer
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
    Chapter:
      type: object
      required: [id, subjectId, number, title, status, mastery]
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        number:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, complete, marked_for_review]
        mastery:
          type: integer
        topics:
          type: array
          items:
            $ref: "#/components/schemas/Topic"
    Topic:
      type: object
      required: [id, chapterId, title, status]
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, complete, needs_review]
        isImportant:
          type: boolean
    Progress:
      type: object
      required: [id, studentId, subjectId, chapterId, topicId, status, updatedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        subjectId:
          type: integer
        chapterId:
          type: integer
        topicId:
          type: ["integer", "null"]
        status:
          type: string
        updatedAt:
          type: string
          format: date-time
    ProgressInput:
      type: object
      required: [topicId, status]
      properties:
        topicId:
          type: integer
        status:
          type: string
          enum: [not_started, in_progress, complete, needs_review]
...
    TestResult:
      type: object
      required: [id, testId, testTitle, subjectName, classLevel, totalQuestions, attempted, correct, incorrect, score, mastery, completedAt]
      properties:
        id:
          type: integer
        testId:
          type: integer
        testTitle:
          type: string
        subjectName:
          type: string
        classLevel:
          type: integer
        totalQuestions:
          type: integer
        attempted:
          type: integer
        correct:
          type: integer
        incorrect:
          type: integer
        score:
          type: integer
        mastery:
          type: integer
        completedAt:
          type: string
          format: date-time
        wrongAnswers:
          type: array
          items:
            $ref: "#/components/schemas/WrongAnswer"
        topicBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/TopicScore"
    WrongAnswer:
      type: object
      required: [questionId, questionText, yourAnswer, correctAnswer]
      properties:
        questionId:
          type: integer
        questionText:
          type: string
        yourAnswer:
          type: string
        correctAnswer:
          type: string
    TopicScore:
      type: object
      required: [topic, score]
      properties:
        topic:
          type: string
        score:
          type: integer
...
    Bookmark:
      type: object
      required: [id, type, referenceId, title, subject, chapter, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [question, topic, chapter]
        referenceId:
          type: integer
        title:
          type: string
        subject:
          type: string
        chapter:
          type: string
        createdAt:
          type: string
          format: date-time
    BookmarkInput:
      type: object
      required: [type, referenceId, title, subject, chapter]
      properties:
        type:
          type: string
          enum: [question, topic, chapter]
        referenceId:
          type: integer
        title:
          type: string
        subject:
          type: string
        chapter:
          type: string
...