# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AuraLearning Elite API
 * OpenAPI spec version: 0.1.0
 */
import type { BookmarkType } from './bookmarkType';

export interface Bookmark {
  id: number;
  type: BookmarkType;
  referenceId: number;
  title: string;
  subject: string;
  chapter: string;
  createdAt: Date;
}