# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AuraLearning Elite API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  email: string;
  classLevel: number;
  points: number;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: Date;
}