# lib/api-zod/src/generated/types/subjectDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AuraLearning Elite API
 * OpenAPI spec version: 0.1.0
 */
import type { Chapter } from './chapter';

export interface SubjectDetail {
  id: number;
  name: string;
  classLevel: number;
  icon: string;
  color: string;
  mastery: number;
  weeklyGoal: number;
  weeklyCompleted: number;
  chapters: Chapter[];
}