# lib/api-zod/src/generated/types/topic.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AuraLearning Elite API
 * OpenAPI spec version: 0.1.0
 */
import type { TopicStatus } from './topicStatus';

export interface Topic {
  id: number;
  chapterId: number;
  title: string;
  status: TopicStatus;
  isImportant?: boolean;
}